pyproject.toml
[project]
name = "dendr"
version = "0.1.0"
description = "Personal knowledge compiler: Obsidian Daily Notes → LLM-maintained wiki"
requires-python = ">=3.11"
dependencies = [
    "click>=8.1",
    "watchdog>=4.0",
    "pydantic>=2.0",
    "pymupdf>=1.24",
    "llama-cpp-python>=0.3",
    "sqlite-vec>=0.1",
    "numpy>=1.26",
    "fastapi>=0.115",
    "uvicorn>=0.32",
    "httpx>=0.27",
    "ulid-py>=1.1",
]

[project.scripts]
dendr = "dendr.cli:main"

[build-system]
requires = ["setuptools>=75"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]
